requirements-api.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-servicebus==7.11.4
azure-identity==1.15.0
python-multipart==0.0.6
